Background blur optimizations (followup of 20933909)
Instead of rendering to an 11-11-10 (in rec2020 space) and keeping a mask, we have the background blur write directly to the swapchain during the up-sample pass.
The 11-11-10 RT is kept but only for the background blur. BlitUIToHDRPS will take care of moving / blending in rec2020 and will output to the 11-11-10 in rec2020, which will be blurred just like in  20933909.
UpsampleMain will then take care of going back to the swapchain color space / format (and still erase the UI in the same region).
A copy of the swapchain will still happen, but right before UI composition. I will then remove that copy in a next CL for platforms that support UAV flag for swapchain.

Fix UpdateScissorRect that would reset rendertarget1 when being called. Use instead the current FRHIRenderPassInfo to keep the same RT layout when updating the stencilbuffer

#preflight 630c62e9660db81edb63918c
#jira UE-81932
#rb daren.cheng

[CL 21670466 by benjamin rouveyrol in ue5-release-engine-staging branch]