Sequencer: Fixed migrating entities sometimes resulting in component values getting lost
FEntityManager::MigrateEntity is responsible for moving the components of an entity between allocations with mismatching combinations of components (ie, if an entity containing ABCDE has B and D removed, it migrates to an allocation containing ACE). The algorithm for this iterates all the destination component types and relocates component data from the old allocation to the new one, for any components that exist in both. For efficiency, the hole that is left in the old allocation is filled with the last entity in that allocation, the same way TArray::RemoveAtSwap works. For example, if entity ID 3 is migrated, we should see the following result:

Before migration:
[ 0, 1, 2, 3, 4, 5 ]

After migration:
[ 0, 1, 2, 5, 4  ]

Unfortunately there were a few flaws in the way this was implemented. Firstly, any components that existed in the old allocation that do not exist in the new allocation were not correctly relocated when performing the RemoveAtSwap. This meant that (in the example above) entity 5 would get entity 3's values for those components. Secondly there were some errors in the way we relocated component values. Relocation _must_ be performed on uninitialized memory to be correct for RAII, and the memory that is left behind _must not_ be destructed. These constraints were not adhered to because AddEntityToAllocation would always default construct its components, and we would not destruct the previous values for components that were not migrated.

I've tried to clarify this code by encapsulating the RemoveAtSwap logic that destructs a component and relocates the last element in a lambda, and this is now called for all source headers that do not get migrated to the new allocation. For any components that _do_ get migrated, we simply relocate the component and perform the swap manually - no destruction is necessary in this case because both values are kept alive.

#preflight 631b614e2b7fe03eb6fbb135
#rb Max.Chen, Ludovic.Chabant

[CL 22020179 by andrew rodham in ue5-release-engine-staging branch]