Fixed stomping of PrevTimestamp's lower bits.
The size of the segment of aux-data to memcpy into the buffer incorrectly included the size of the aux-header. This would result in writing aux-data off the end of the overflow and into the low-order bits of the timestamp bias. This would inadvertantly cause incorrect, sometimes negative, timestamps for scoped events.

#rb im
#jira UE-163769
#rnx
#preflight 632325d3d5013ce29f43854c

[CL 22028522 by martin ridgers in ue5-release-engine-staging branch]