Fixed runtimes not starting up with certain combinations of trace channels enabled
The redirect buffer was a little too small to contain the callstack that may get traced. Bumping the size isn't the ideal solution but suitable for the now. Alternatives would be to delta compress callstack events (only an option on monolithic builds), take smaller backtraces for trace-related allocations, eliminate redundancy in addresses (i.e. a 32 bit encoding scheme), or possibly segmented deduplication.

#rb im
#jira UE-163770
#rnx
#preflight 63242bb274361465144c7c4d

[CL 22047209 by martin ridgers in ue5-release-engine-staging branch]