Removed event redirection when allocating trace memory
It is only when allocating buffers to store normal trace events that care needs to be taken to capture events and blit them to the appropriate place. This is already in place in Writer_NextBuffer(). This also handles cases where traced events are artifacts of Trace allocating memory for other purposes. So the redirect removed by this change was incorrect. Under the correct conditions, it would accelerate events ahead of others due to the SendData() call and result in the incorrect ordering events being sent across the wire.

#rb im
#jira UE-163770
#rnx
#preflight 63296e009840225da24c032e

[CL 22089724 by martin ridgers in ue5-release-engine-staging branch]