Fixed TraceAuxilary starting Trace's thread too early.
Trace's thread can't be started before Trace has been initialised as some internals are not ready prior to that (i.e. the cache and the tail). TraceAux attempts to control the worker thread and indirectly started it too early when processing command line arguments. This would cause data races between the worker thread and initialisation.

#rb im
#rnx
#jira none
#preflight 63296dab0a4ad044de754095

[CL 22089725 by martin ridgers in ue5-release-engine-staging branch]