Core - Introducing FindPackageReferencesInObject, this utility function provides a pretty path from the root object to any references to any of the packages provided.
Given a root object, this function will call the callback every time it encounters a reference in the root object (including all inlined/instanced UObjects.  The returned references will be in an easily understood debugging format so that they're easy to understand the context.

For example, Actions[2](GameFeatureAction_Activities)->ObjectActivitiesInfo[0].TargetReference

Where TargetReference was one of the listed PackagesReferenced.  The point of this function is to assist in the reporting of references between objects you need to show the developer.

[CL 22356759 by nick darnell in ue5-release-engine-staging branch]