[PCG] Localized node tooltips. Demonstrated for projection node.
If a tooltip is specified, it will show this, otherwise it will just show the node name. I saw this behaviour on a couple of blueprint nodes so went for this.

Pin tooltips are something different, not covered by this change. I think there is utility in pin tooltips but I did not figure out how to get these stood up in my initial time box and they might take more work. I could look at them in a follow up CL if we want that too.

#preflight 6369231ddc30a4ce964325b3
#rb julien.lheureux, adrien.logut

[CL 23026115 by huw bowles in ue5-release-engine-staging branch]