Allow backtrace captures to vary in size on a per-platform basis.
Some platforms have threads with very small stacks that we don't appear to be able to control. This places a limit on how much stack is available to capture a callstack in situations where memory hooks are in place. Curl's host name resolving threads an the CurlCalloc() hooks are one such example.

#rb cm
#rnx
#preflight 636d00ad376a9cd6a8f4e22c
#ushell-cherrypick of 23076081 by Martin.Ridgers

[CL 23099496 by martin ridgers in ue5-release-engine-staging branch]