UBT: Edit and Continue: Unified the property names used to enable edit and continue.
On some platforms we used bUseLiveCodeEdit, which was expected to be present in the per-platform configuration block. At the same time, on Windows we rely on bSupportEditAndContinue from BuildConfiguration.

Modified the console platforms to use BuildConfiguration.bSupportEditAndContinue for consistency. However, there is some value in being able to control this setting per-platform, so I've left the existing properties and just renamed them bUseLiveCodeEdit -> bSupportEditAndContinue.

As a result, all platforms that support Edit and Continue are affected by BuildConfiguration.bSupportEditAndContinue. However, if XXPlatform.bSupportEditAndContinue is presetnt, it overrides the global.

Removed ClangToolChainOptions.EnableLiveCodeEditing and related code as it turns out it wasn't really used.

#preflight 637274a99e3bea8079f98af5
#rb Joe.Kirchoff

[CL 23120415 by wojciech krywult in ue5-release-engine-staging branch]