Fix Editor crash when updating volume texture source from a function.
Note: Previously the source was first initialized with a dummy texture in order to determine the number of bytes per pixel. However the dummy texture was created with zero size and non-zero mip count, which was incorrect. Creating a dummy texture is no longer necessary, as we now have a static function to determine bytes per pixel. In addition to that, the referenced source texture is now reset when generating the volume source from a function.

#jira UE-169133
[FYI] charles.bloom
#preflight 6373f9ae232e3d12cbe0ade5

[CL 23142322 by alexander suvorov in ue5-release-engine-staging branch]