Sanitise received payload size to within sensible limits.
Any random data that might be received by the CBOR server could cause it to allocate a lot of memory, or suffer a signed overflow and collapse. This can happen for example if a machine is port scanned by an IT department. To guard against this, the input is checked to ensure it is within a suitable before use.

#rb im
#rnx
#preflight none

[CL 23246369 by martin ridgers in ue5-release-engine-staging branch]