Only create traces with the store when valid trace data has been detected.
Prior to this change, a trace was registered with the store component before creating the relay that receives data over the socket and writes it to a file. While the relay does check the preamble of the data received, and rejects anything that is not a trace, it does not clean up the store. Now no trace is created until the relay has determined it has a trace to receive data for. This stops zero-sized traces littering a users disk which can happen if the machine is port scanned by their IT department.

#rb im
#rnx
#preflight none

[CL 23246570 by martin ridgers in ue5-release-engine-staging branch]