[PCG] Fix unnecessary collapse of spline projected onto landscape.
In this case we can implement a SamplePoint to see if something is in the projection (provided the direction is straight down, things gets murky-er for projections in other directions such as a shallow angle onto the terrain which results in 'shadows' - it would take more work & computation to decide if a point is in the projection or in a shadow).

Also fixes bug in the Project() function that derived the projection axis from the projected point coordinates by looking for smallest coordinate. As an example, the origin (0,0,0) would project to (0,0,0) but this does not tell us which axis was used for projection. Use projection direction directly instead which is not ambiguous.

#rb julien.lheureux, adrien.logut
#preflight 6380d4de8b12eb83a7b9f024

[CL 23266611 by huw bowles in ue5-release-engine-staging branch]