Path Tracer: Fix interpretation of clearcoat parameter
Setting clearcoat to 0 was not properly fading out the coating. After careful re-reading of the raster evaluation, the logic was subtly different between the two. I also refactored the code to directly compute the three lobes without falling back to the logic in default lit. This prepares the ability to sample diffuse and specular independently.

Implement additional version of lobe selection for 3 and 4 lobed materials. Like the previous method, these are robust to any combination of lobes having zero weights which helps avoid special cases and numerical issues.

Removed unused variables in ClearCoatBxDF

#rb none
#jira none
[FYI] Sebastien.Hillaire
#preflight 6387eb9bd76995cb14c8ec69

[CL 23344968 by chris kulla in ue5-release-engine-staging branch]