[PCG] Add ProjectPoint function to spatial data which performs a projection.
Often SamplePoint can be expressed as a projection one way or another - a majority of the functionality is shared. Therefore data's that have received this new ProjectPoint function have had their SamplePoint refactored to use it.

This has the benefit of surfacing the fact that the landscape data does actually perform a projection in its SamplePoint (for now - will be addressed next).

UPCGProjectionData now calls ProjectPoint. To stop this from breaking loads of things, UPCGSpatialData has a crutch fallback that forwards the call to SamplePoint to retain behaviour for most data's. The exception is UPCGPointData where I've proceeded to actually purify SamplePoint, which will make cases which wanted projection via SamplePoint() instead of using UPCGProjectionData fail. I will message this.

UPCGProjectionData is updated to use the new project function, and can pass its projection params into it which is nice. Slightly oddly, the project function takes Transform and Bounds from the source point but not the rest of the info like Color or metadata, and therefore it only makes use of some of the projection params (bProjectPositions, bProjectScales, bProjectRotations) and not the rest. I feel this is odd but acceptable - I think project should only need Transform and Bounds as input rather than an entire point. If a caller has access to more data then it can decide which data to take using the other projection params which is exactly what UPCGProjectionData does.

Some decisions will need to be made about how projection is implemented in some of the data's before we can move the crutch fallback. TODO's are added to mark some non-trivial cases.

#preflight 638dc8399242c01326016e32
#rb julien.lheureux, adrien.logut
#preflight none

[CL 23388727 by huw bowles in ue5-release-engine-staging branch]