Fix iOS GameCenter login flow on iOS to fix Lyra issues
Setting GKLocalPlayer authenticateHandler only shows a native login UI/greeting the first time it is set. After that it receives GameCenter login state updates. Previously looks like it was assuming a login flow was initiated from that. Once GameCenter reports an authentication failure error no more errors are notified through the handler so it was possible that OnLoginComplete delegates were not called after a first failure anymore

It is possible tha a user moves the app to background and changes the GameCenter user from device settings, so this changes also adds triggering OnLoginStatusChanged delegates that were missing

[REVIEW] [at]sam.zamani, [at]bertrand.carre, [at]chris.varnsverry, [at]ben.zeigler
#jira UE-168605
#preflight

[CL 23426327 by rafa lecina in ue5-release-engine-staging branch]