Lumen - replace cone skylight evaluation with a ray sample in SWRT radiosity.
There's no point in having cone evaluation only for a single place in the pipeline. Additionally this was causing a lot of leaking with bright skylights. E.g. when a ray hits black lower sky hemisphere it still could bring in color from the upper one due to the ray footprint.

[FYI] Daniel.Wright
#preflight skip

[CL 23446617 by krzysztof narkowicz in ue5-release-engine-staging branch]