[PCG] Fix for Composite type - it was wrong to have it on its own isolated bit, because Concrete "is a" Composite, ie Composite was supposed to be a superset of Concrete.
It may be quirky that there is a new type bit ShapeBoolean that is not directly used in the wider code. I.e. Union still outputs Composite. I believe this is correct - a Union can create a boolean if it has multiple inputs, or if it only has one input it passes it through in which case it can produce Concrete. So it can output concrete or boolean and its type should stay at Composite. Therefore the ShapeBoolean bit is only there to differentiate type but is not needed in the wider code.

#rb julien.lheureux, thomas.tedemalm
#preflight 639345950d013d47ef1aad2e

[CL 23463314 by huw bowles in ue5-release-engine-staging branch]