Path Tracer: Decouple refraction IOR from specularity for solid and thin glass models
In the thin case, Ior now affects the transmitted roughness, so that rough refraction is progressively disabled as Ior is reduced toward 1.0. When not using refraction, ThinTranslucent materials are rendered using transparency, while with refraction they are rendered with an actual bounce (to allow for roughness, and for consistency with the solid case). SpecularColor controls the fresnel effect.

In the solid case, Ior now only affects ray bending, while SpecularColor controls the fresnel effect. Refraction must be enabled in the material to get a refractive event. Disabling refraction means the surface is simply transparent.

#rb Sebastien.Hillaire
#jira UE-172456
#preflight 639309ca9549ddaa288c01ed

[CL 23464398 by chris kulla in ue5-release-engine-staging branch]