[PCG] Split surface sampler node pins into a Surface input and a Bounding Shape input.
For a description of the new behaviour, see the node and pin tooltips in the code.

Deprecation code connects the current 'In' input to the Bounding Shape which should preserve current graphs.

A new API is added to SpatialData that looks for a shape of a given dimension. This happened implicitly before but is explicit now. This could be extended in the future to return multiple shapes if there are bugs with it picking the first encountered. The implementation of this is fairly simple and seems to work in my cases, but notably it does not make use of some previous things like HasNonTrivialTransform(), if this sounds like trouble I could use a chat/test case that exhibits issues or limitations.

#rb julien.lheureux
#preflight 6399934a0a67152550c5a777

[CL 23524523 by huw bowles in ue5-release-engine-staging branch]