Thin-translucent: adjust logic for thin-translucent to be a bit more consistent with the path tracer.
The fresnel calculations were needlessly expensive and can be simplified as entering and exiting events must have the same fresnel factor. Added the mathematica "proof" of this in the comments.

The specular color was only using Spec and not Metallic. This limited the effective IOR range of thin translucent materials to 1.8 at most. Now metallic is no longer scaled by opacity, which allows for creating surfaces made of higher IORs like diamond for example.

This is motivated by the fact that Strata will be using Diffuse/Specular workflow internally and therefore does not treat metallic specially in this case.

Adjust RT Translucency to match the path tracer behavior, and implement the missing blend modes.

#jira UE-172456
#rb Sebastien.Hillaire
#preflight 639912de2540a78d274ccef6

[CL 23563361 by chris kulla in ue5-release-engine-staging branch]