AssetDefinition - Converting the Haptic AssetTypeActions to AssetDefinitions.
The copying localization ID is now an option stated as such instead of a right click showing the ID you click.  In order to show the ID the asset has to be loaded, so instead it's now a copy operation that then loads the asset before copying the ID.

Fixing a regression that stopped showing export options for unloaded assets.  The code was using FastGetAsset, which won't load the asset if unloaded.  The new code just looks at the package flags from the FAssetData, no need to attempt to load the asset.

#jira UE-165574
#jira UE-173203

[CL 23587423 by nick darnell in ue5-release-engine-staging branch]