nDisplay: iteration over world actors looking for hidden primitives now only happens when not in game.
As a consequence, it will now not render visualization components. That was a hack added in the past to support Text3d, but Text3d will be updated to avoid this issue instead.

Since the iteration is also needed for in-editor preview, parallelized it for higher performance. For CitySampleVP it went from 18 ms to 2 ms.

Added a missing member variable initialization.

Added a convenience function to programmatically freeze/unfreeze the outers.

#jira
#rb vitalii.boiko
#preflight 63b9f88b47321b9d89b9ac98

[CL 23608143 by alejandro arango in ue5-release-engine-staging branch]